feat: create TokenManager class
