+++ src/pages/api/member/[shortUrl].ts
@@ -1,7 +1,7 @@
 import { ResponseContent } from "@/model/Common";
 import sequelize, { UrlSequelize } from "@/server/db";
 import { getUrlWithOgByAccount, URLJoinObContent } from "@/server/method";
-import { getAccountFromToken, client, connectToRedis } from "@/util/decode";
+import TokenManager from "@/util/token";
 import { NextApiRequest, NextApiResponse } from "next";
 
 interface MemberResponseContent extends ResponseContent {
@@ -15,8 +15,7 @@ export default async function handler(
   try {
     await sequelize.authenticate();
 
-    await connectToRedis(client);
-    const accountFromToken = await getAccountFromToken(req, client);
+    const accountFromToken = await TokenManager.getAccountFromToken(req);
 
     if (accountFromToken) {
       if (req.method !== "DELETE") {
